Renamed is-test-mode to test-modes
