+++ src/api/me/index.ts
@@ -6,7 +6,7 @@
 
 import { UserResponse } from '../users/types'
 import { defaultFetchConfig, expectResponseCode, getApiUrl } from '../utils'
-import { isMockMode } from '../../utils/is-test-mode'
+import { isMockMode } from '../../utils/test-modes'
 
 export const getMe = async (): Promise<UserResponse> => {
   const response = await fetch(getApiUrl() + `/me${ isMockMode() ? '-get' : '' }`, {
